requirements.txt
aiohttp==3.7.4.post0
async-irc==0.1.7
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
chardet==4.0.0
httplib2==0.19.1
idna==3.1
idna-ssl==1.1.0
multidict==5.1.0
pkg-resources==0.0.0
py-irclib==0.3.0
pyparsing==2.4.7
PyYAML==5.4.1
typing-extensions==3.7.4.3
yarl==1.6.3
